+++ src/main/java/division/DivisionInterface.java
@@ -1,6 +1,6 @@
 package division;
 
 public interface DivisionInterface {
-    double division(int one, int two) throws ZeroException;
+    int division(int one, int two) throws ZeroException;
     double division(double one, double two) throws ZeroException;
 }
